Move marshalling code to its own file
